Refactor new page as an edit task component
